create class for creations and modules for menu, run and options
