+++ app.rb
@@ -2,9 +2,11 @@ require_relative './student'
 require_relative './teacher'
 require_relative './rental'
 require_relative './book'
+require_relative './show_menu'
+require_relative './creation'
 
-# rubocop:disable Metrics/ClassLength
 class App
+  include ShowMenu
   attr_reader :book, :people, :rental
 
   def initialize
@@ -26,42 +28,43 @@ class App
     end
   end
 
-  def library_menu
-    puts ''
-    puts ''
-    puts 'Welcome to the School Library App!'
-    puts ''
-    puts ''
-    puts 'Please choose one of these options by entering a number'
-    puts ''
-    puts ''
-    puts '1- List all available books'
-    puts '2- List all people'
-    puts '3- Create a person'
-    puts '4- Create a book'
-    puts '5- Create a rental'
-    puts '6- List all rentals for a given person(ID)'
-    puts '7- Exit'
-  end
+  # def library_menu
+  #   puts ''
+  #   puts ''
+  #   puts 'Welcome to the School Library App!'
+  #   puts ''
+  #   puts ''
+  #   puts 'Please choose one of these options by entering a number'
+  #   puts ''
+  #   puts ''
+  #   puts '1- List all available books'
+  #   puts '2- List all people'
+  #   puts '3- Create a person'
+  #   puts '4- Create a book'
+  #   puts '5- Create a rental'
+  #   puts '6- List all rentals for a given person(ID)'
+  #   puts '7- Exit'
+  # end
 
-  def options(input)
-    case input
-    when '1'
-      list_books
-    when '2'
-      list_people
-    when '3'
-      create_person
-    when '4'
-      create_book
-    when '5'
-      create_rental
-    when '6'
-      list_rentals
-    else
-      'Enter a number between from 1 to 7'
-    end
-  end
+  # def options(input)
+  #   creation = Creation.new(@people, @book, @rental)
+  #   case input
+  #   when '1'
+  #     list_books
+  #   when '2'
+  #     list_people
+  #   when '3'
+  #     @people.push(creation.create_person)
+  #   when '4'
+  #     @book.push(creation.create_book)
+  #   when '5'
+  #     @rentals.push(creation.create_rental)
+  #   when '6'
+  #     list_rentals
+  #   else
+  #     'Enter a number between from 1 to 7'
+  #   end
+  # end
 
   def list_books
     if @book.empty?
@@ -89,47 +92,85 @@ class App
     show_menu
   end
 
-  def create_person
-    print 'Do you want to create a student (1) or a teacher (2)? ..Please input number:  '
-    option = gets.chomp.to_i
+  # class Creation
+  #   include ShowMenu
+  #   def initialize(people = [], book = [], rental = [])
+  #     @book = book
+  #     @people = people
+  #     @rental = rental
+  #   end
 
-    case option
-    when 1
-      create_student
-    when 2
-      create_teacher
-    else
-      puts 'Please select 1 or 2 !'
-    end
-  end
+  #   def create_person
+  #     print 'Do you want to create a student (1) or a teacher (2)? ..Please input number:  '
+  #     option = gets.chomp.to_i
 
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
-    book = Book.new(title, author)
-    @book.push(book)
-    puts 'Book successfully created!'
-  end
+  #     case option
+  #     when 1
+  #       create_student
+  #     when 2
+  #       create_teacher
+  #     else
+  #       puts 'Please select 1 or 2 !'
+  #     end
+  #   end
 
-  def create_rental
-    puts 'Select a book from the list (Enter number)'
-    @book.each_with_index { |item, i| puts "#{i}-- #{item.title}, #{item.author}" }
-    index = gets.chomp.to_i
-    book = @book[index]
-    puts 'Select a person from the list (Enter number)'
-    @people.each_with_index { |item, i| puts "#{i}- #{item.name}" }
-    person_index = gets.chomp.to_i
-    person = @people[person_index]
-    print 'Date(yyyy/mm/dd): '
-    date = gets.chomp
-    rental = Rental.new(date, book, person)
-    puts rental
-    @rental.push(rental)
-    puts 'Rental created successfully'
-    show_menu
-  end
+  #   def create_book
+  #     print 'Title: '
+  #     title = gets.chomp
+  #     print 'Author: '
+  #     author = gets.chomp
+  #     book = Book.new(title, author)
+  #     @book.push(book)
+  #     puts 'Book successfully created!'
+  #     show_menu
+  #   end
+
+  #   def create_rental
+  #     puts 'Select a book from the list (Enter number)'
+  #     @book.each_with_index { |item, i| puts "#{i}-- #{item.title}, #{item.author}" }
+  #     index = gets.chomp.to_i
+  #     book = @book[index]
+  #     puts 'Select a person from the list (Enter number)'
+  #     @people.each_with_index { |item, i| puts "#{i}- #{item.name}" }
+  #     person_index = gets.chomp.to_i
+  #     person = @people[person_index]
+  #     print 'Date(yyyy/mm/dd): '
+  #     date = gets.chomp
+  #     rental = Rental.new(date, book, person)
+  #     puts rental
+  #     @rental.push(rental)
+  #     puts 'Rental created successfully'
+  #     show_menu
+  #   end
+
+  #   private
+
+  #   def create_student
+  #     print 'Age: '
+  #     age = gets.chomp.to_i
+  #     print 'Name: '
+  #     name = gets.chomp
+  #     print 'Has parent permission? [Y/N]'
+  #     permission = gets.chomp
+  #     new_student = Student.new(age, permission, name)
+  #     @people.push(new_student)
+  #     puts 'Student created successfully'
+  #     show_menu
+  #   end
+
+  #   def create_teacher
+  #     print 'Age: '
+  #     age = gets.chomp.to_i
+  #     print 'Name: '
+  #     name = gets.chomp
+  #     print 'Specialization: '
+  #     specialization = gets.chomp
+  #     teacher = Teacher.new(age, specialization, name)
+  #     @people.push(teacher)
+  #     puts 'Teacher created successfullly'
+  #     show_menu
+  #   end
+  # end
 
   def list_rentals
     puts 'Select ID of any person (Please type the number of the ID)'
@@ -142,40 +183,11 @@ class App
     show_menu
   end
 
-  def show_menu
-    puts ''
-    print 'Press Enter to go back to menu'
-    gets.chomp
-    puts ''
-    run
-  end
-
-  private
-
-  def create_student
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Has parent permission? [Y/N]'
-    permission = gets.chomp
-    new_student = Student.new(age, permission, name)
-    @people.push(new_student)
-    puts 'Student created successfully'
-    show_menu
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Specialization: '
-    specialization = gets.chomp
-    teacher = Teacher.new(age, specialization, name)
-    @people.push(teacher)
-    puts 'Teacher created successfullly'
-    show_menu
-  end
+  # def show_menu
+  #   puts ''
+  #   print 'Press Enter to go back to menu'
+  #   gets.chomp
+  #   puts ''
+  #   run
+  # end
 end
-# rubocop:enable Metrics/ClassLength
