+++ library_menu.rb
@@ -1,0 +1,19 @@
+module LibraryMenu
+  def library_menu
+    puts ''
+    puts ''
+    puts 'Welcome to the School Library App!'
+    puts ''
+    puts ''
+    puts 'Please choose one of these options by entering a number'
+    puts ''
+    puts ''
+    puts '1- List all available books'
+    puts '2- List all people'
+    puts '3- Create a person'
+    puts '4- Create a book'
+    puts '5- Create a rental'
+    puts '6- List all rentals for a given person(ID)'
+    puts '7- Exit'
+  end
+end
